# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Space Traffic API
  version: "1.1"
  description: |
    Simulated solar-system shipping traffic API for data-engineering demos.

    This API is fully public and does not require authentication.
servers:
  - url: http://localhost:8000
    description: Local development server
components:
  parameters:
    Limit1000:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Limit5000:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 5000
        default: 1000
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Order:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
    SinceId:
      name: since_id
      in: query
      schema:
        type: integer
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: unknown scenario 'blastoff'
...
    Station:
      type: object
      required: [id, name, body_name, body_type, parent_body, allowed_size_classes]
      properties:
        id:
          type: string
          example: STN-PLANET-EARTH
        name:
          type: string
          example: Avalon Prime Port
        body_name:
          type: string
          example: Earth
        body_type:
          type: string
          enum: [planet, moon, asteroid]
        parent_body:
          type: string
          example: Sun
        allowed_size_classes:
          type: array
          items:
            type: string
            enum: [small, medium, large, xlarge]
          example: [small, medium]
    Ship:
      type: object
      required: [id, name, faction, ship_type, size_class, displacement_million_m3, home_station_id, captain_name, cargo]
      properties:
        id:
          type: string
          example: SHIP-00042
        name:
          type: string
          example: Iron Pioneer
        faction:
          type: string
          enum: [merchant, government, military, bounty_hunter]
        ship_type:
          type: string
          example: Freighter
        size_class:
          type: string
          enum: [small, medium, large, xlarge]
        displacement_million_m3:
          type: number
          format: float
          example: 1.4
        home_station_id:
          type: string
          example: STN-MOON-TITAN
        captain_name:
          type: string
          example: Mira Voss
        cargo:
          type: string
          example: titanium_alloy_billets
    ShipState:
      type: object
      required: [ship_id, status, in_transit, ship_age_days, updated_at]
      properties:
        ship_id:
          type: string
          example: SHIP-00042
        status:
          type: string
          enum: [active, decommissioned, destroyed]
        current_station_id:
          type: [string, "null"]
          example: STN-PLANET-MARS
        in_transit:
          type: integer
          enum: [0, 1]
        source_station_id:
          type: [string, "null"]
          example: STN-MOON-TITAN
        destination_station_id:
          type: [string, "null"]
          example: STN-PLANET-EARTH
        departure_time:
          type: [string, "null"]
          format: date-time
          description: Simulated timeline departure timestamp.
        departure_time_simulated:
          type: [string, "null"]
          format: date-time
          description: Explicit alias for simulated departure timestamp.
        est_arrival_time:
          type: [string, "null"]
          format: date-time
          description: Simulated timeline estimated arrival timestamp sampled at departure creation time.
        est_arrival_time_simulated:
          type: [string, "null"]
          format: date-time
          description: Explicit alias for simulated estimated arrival timestamp.
        ship_age_days:
          type: number
          format: float
          example: 1482.5
        updated_at:
          type: string
          format: date-time
          description: Runtime update/persistence timestamp (transitional until explicit wall-clock split lands).
        updated_at_simulated:
          type: string
          format: date-time
          description: Simulated timeline update timestamp.
        updated_at_wall:
          type: [string, "null"]
          format: date-time
          description: Wall-clock timestamp when this state row was last persisted.
    Departure:
      type: object
      required: [id, event_uid, departure_time, malformed, fault_flags]
      properties:
        id:
          type: integer
          description: Auto-incrementing row ID. Use as cursor for polling.
          example: 1024
        event_uid:
          type: string
          example: EVT-20150101-0042-SHIP-00042
        departure_time:
          type: string
          format: date-time
          description: Simulated timeline departure timestamp.
        departure_time_simulated:
          type: string
          format: date-time
          description: Explicit alias for simulated departure timestamp.
        ship_id:
          type: [string, "null"]
          example: SHIP-00042
        source_station_id:
          type: [string, "null"]
          example: STN-MOON-TITAN
        destination_station_id:
          type: [string, "null"]
          example: STN-PLANET-EARTH
        est_arrival_time:
          type: [string, "null"]
          format: date-time
          description: Simulated timeline estimated arrival timestamp.
        est_arrival_time_simulated:
          type: [string, "null"]
          format: date-time
          description: Explicit alias for simulated estimated arrival timestamp.
        recorded_at_wall:
          type: [string, "null"]
          format: date-time
          description: Wall-clock timestamp when the departure row was persisted.
        scenario:
          type: [string, "null"]
          description: Active scenario name at time of departure, or null.
          example: war
        fault_flags:
          type: array
          items:
            type: string
          description: List of active fault types that affected this event.
          example: []
        malformed:
          type: boolean
          description: True if the payload field was intentionally corrupted by fault injection.
        payload:
          description: Full structured departure payload, or null/corrupt string if malformed.
    ControlEvent:
      type: object
      required: [id, event_time, event_type, action, payload]
      properties:
        id:
          type: integer
          description: Auto-incrementing row ID. Use as cursor for polling.
          example: 7
        event_time:
          type: string
          format: date-time
          description: Event timestamp used by current runtime path (explicit wall-clock/simulated split planned).
        event_time_simulated:
          type: string
          format: date-time
          description: Explicit simulated event timestamp.
        recorded_at_wall:
          type: [string, "null"]
          format: date-time
          description: Wall-clock timestamp when the control event row was persisted.
        event_type:
          type: string
          enum: [control, scenario, fault, lifecycle, pirate]
          example: scenario
        action:
          type: string
          example: activated
        payload:
          type: object
          description: Event-specific details (scenario config, fault names, reset seed, etc.).
    ActiveScenario:
      type: [object, "null"]
      properties:
        name:
          type: string
          enum: [war, shortage, solar_flare]
        intensity:
          type: number
          format: float
          minimum: 0.0
          maximum: 5.0
          example: 1.5
        duration_seconds:
          type: integer
          example: 300
        scope:
          type: object
          properties:
            type:
              type: string
              enum: [global, region]
        started_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
    FaultEntry:
      type: object
      required: [rate, scope, duration_seconds, started_at]
      properties:
        rate:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          example: 0.2
        scope:
          type: object
        duration_seconds:
          type: integer
          example: 300
        started_at:
          type: string
          format: date-time
        ends_at:
          type: [string, "null"]
          format: date-time
    PaginatedMeta:
      type: object
      required: [count, total_count, offset, limit]
      properties:
        count:
          type: integer
          description: Number of items in this response.
        total_count:
          type: integer
          description: Total matching items across all pages.
        offset:
          type: integer
        limit:
          type: integer
paths:
...
  /stations:
    get:
      summary: List stations
      parameters:
        - name: body_type
          in: query
          schema:
            type: string
            enum: [planet, moon, asteroid]
        - $ref: "#/components/parameters/Limit5000"
        - $ref: "#/components/parameters/Offset"
        - name: order_by
          in: query
          schema:
            type: string
            enum: [id, name, body_name, body_type, parent_body]
            default: body_type
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          description: Paginated station list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedMeta"
                  - type: object
                    required: [stations]
                    properties:
                      stations:
                        type: array
                        items:
                          $ref: "#/components/schemas/Station"
  /ships:
    get:
      summary: List ships
      parameters:
        - name: faction
          in: query
          schema:
            type: string
            enum: [merchant, government, military, bounty_hunter]
        - name: home_station_id
          in: query
          schema:
            type: string
        - name: cargo
          in: query
          schema:
            type: string
        - name: ship_type
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Limit5000"
        - $ref: "#/components/parameters/Offset"
        - name: order_by
          in: query
          schema:
            type: string
            enum: [id, name, faction, ship_type, cargo, home_station_id, size_class]
            default: id
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          description: Paginated ship list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedMeta"
                  - type: object
                    required: [ships]
                    properties:
                      ships:
                        type: array
                        items:
                          $ref: "#/components/schemas/Ship"
  /ships/state:
    get:
      summary: List live ship states
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, decommissioned, destroyed]
        - name: in_transit
          in: query
          schema:
            type: string
            enum: ["0", "1", "true", "false", "yes", "no", "on", "off"]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5000
            default: 500
      responses:
        "200":
          description: Ship state list
          content:
            application/json:
              schema:
                type: object
                required: [ships, count]
                properties:
                  ships:
                    type: array
                    items:
                      $ref: "#/components/schemas/ShipState"
                  count:
                    type: integer
...
  /departures:
    get:
      summary: Poll departure events
      parameters:
        - $ref: "#/components/parameters/SinceId"
        - name: since_time
          in: query
          schema:
            type: string
            format: date-time
        - $ref: "#/components/parameters/Limit1000"
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          description: Departure events page
          content:
            application/json:
              schema:
                type: object
                required: [departures, count, next_since_id]
                properties:
                  departures:
                    type: array
                    items:
                      $ref: "#/components/schemas/Departure"
                  count:
                    type: integer
                  next_since_id:
                    type: [integer, "null"]
                    description: Pass as since_id on your next call to get only new events.
...
  /control-events:
    get:
      summary: Poll control-plane events
      parameters:
        - $ref: "#/components/parameters/SinceId"
        - $ref: "#/components/parameters/Limit1000"
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          description: Control events page
          content:
            application/json:
              schema:
                type: object
                required: [control_events, count, next_since_id]
                properties:
                  control_events:
                    type: array
                    items:
                      $ref: "#/components/schemas/ControlEvent"
                  count:
                    type: integer
                  next_since_id:
                    type: [integer, "null"]
...
  /config:
    get:
      summary: Get runtime snapshot
      responses:
        "200":
          description: Runtime configuration and active scenario/fault state
          content:
            application/json:
              schema:
                type: object
                properties:
                  base_min_events_per_minute:
                    type: number
                  base_max_events_per_minute:
                    type: number
                  deterministic_mode:
                    type: boolean
                  deterministic_seed:
                    type: [integer, "null"]
                  deterministic_start_time:
                    type: string
                    format: date-time
                    description: Seeded simulated start timestamp (used in deterministic mode).
                  simulation_now:
                    type: string
                    format: date-time
                    description: Current simulated timeline timestamp.
                  simulation_time_scale:
                    type: number
                    minimum: 0.1
                    description: Sim-to-wall ratio semantics (higher means faster simulated progression).
                  retention_max_rows:
                    type: integer
                  db_max_size_mb:
                    type: integer
                  active_scenario:
                    $ref: "#/components/schemas/ActiveScenario"
                  active_faults:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/FaultEntry"
                  pirate_event:
                    type: object
                    properties:
                      active:
                        type: boolean
                      anchor_body:
                        type: [string, "null"]
                      strength:
                        type: number
                      started_at:
                        type: [string, "null"]
                        format: date-time
                      affected_station_ids:
                        type: array
                        items:
                          type: string
                  effective_lifecycle:
                    type: object
                    description: Lifecycle config merged with any active scenario overrides.
                  effective_ship_generation:
                    type: object
                    description: Ship generation config merged with any active scenario overrides.
    patch:
      summary: Patch runtime configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deterministic_mode:
                  type: boolean
                deterministic_seed:
                  type: integer
                deterministic_start_time:
                  type: string
                  format: date-time
                  description: Simulated epoch baseline used on startup/reset and when re-anchoring simulation_now.
                simulation_time_scale:
                  type: number
                  minimum: 0.1
                db_max_size_mb:
                  type: integer
                  minimum: 50
      responses:
        "200":
          description: Updated runtime snapshot
          content:
            application/json:
              schema:
                type: object
                description: Updated runtime snapshot (same schema as GET /config).
...
  /scenarios/activate:
    post:
      summary: Activate a scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  enum: [war, shortage, solar_flare]
                intensity:
                  type: number
                  minimum: 0
                  maximum: 5
                  default: 1.0
                duration_seconds:
                  type: integer
                  minimum: 1
                  default: 300
                scope:
                  type: object
      responses:
        "200":
          description: Active scenario set
          content:
            application/json:
              schema:
                type: object
                required: [active_scenario]
                properties:
                  active_scenario:
                    $ref: "#/components/schemas/ActiveScenario"
        "400":
          description: Invalid scenario payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /faults/activate:
    post:
      summary: Activate one or more faults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                faults:
                  type: object
      responses:
        "200":
          description: Active faults updated
          content:
            application/json:
              schema:
                type: object
                required: [active_faults]
                properties:
                  active_faults:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/FaultEntry"
        "400":
          description: Invalid fault payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /faults/deactivate:
    post:
      summary: Deactivate selected (or all) faults
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                names:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Active faults updated
          content:
            application/json:
              schema:
                type: object
                required: [active_faults]
                properties:
                  active_faults:
                    type: object
        "400":
          description: Invalid names payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /control/reset:
    post:
      summary: Reset simulation state
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                seed:
                  type: integer
      responses:
        "200":
          description: Reset complete with runtime snapshot
          content:
            application/json:
              schema:
                type: object
                required: [status, runtime]
                properties:
                  status:
                    type: string
                    example: reset
                  runtime:
                    type: object
                    description: Runtime snapshot after reset.
...